Add files via upload
This is my final report on URL shortener. The purpose of this URL shortener is to provide a way to shorten long URLs into shorter, more manageable URLs. This can be useful for a variety of purposes, such as making it easier to share long URLs on social media or in text messages, tracking the number of clicks on a URL, and preventing malicious URLs from being clicked.